♻️ refactor: 컴포넌트 컨벤션 유지
https://www.notion.so/shineshine/FE-4923eb16229a4e0584d57187a5a80746?pvs=4#c4cce9000a074fe9bc1e3651788a7ea0
