+++ src/app/auth/logout/LogoutComponent.tsx
