--- src/app/auth/logout/components/index.ts
@@ -1,1 +1,0 @@
-export { default as LogoutComponent } from './LogoutComponent';
