+++ src/app/auth/logout/page.tsx
@@ -1,4 +1,4 @@
-import { LogoutComponent } from './components';
+import LogoutComponent from './LogoutComponent';
 
 export default function Page() {
   return <LogoutComponent />;
